The Great Gradle-Resolver Refactor Revolution a.k.a. don't check diffs
Behold, and bear witness to my complete abuse of version control
technology, fashioned like a lovecraftian Ship of Theseus, an
undertaking so vile it lived on 2 other computers without being able to
compile for over a year: a commit that touches every file of the
project! Diff tools beware, even the most advanced of algorithms shan't
discern what remains of the original state of the project. Those in the
future who wish to look at the history of each file shall understand the
glory of today's revolution. Blames will not reach too far backwards to
the past, because today marks a great leap forward. Do not be concerned;
the past isn't being burned or rewritten, but to have any hope of
understanding this project's history, your one source of truth shall be
what the party deems appropriate to write here.

- Switched the project over to use Gradle instead of Eclipse. The
  project can be run with `./gradlew run`. Eclipse support is removed,
  go use vscode or vim or something.
  - This means most of the files were moved from `src/` to
    `src/main/java` as gradle convention
  - Natives and other dependencies are now automatically fetched by
    gradle. Anyone should be able to run this project with minimal
    setup.
- Converted tabs to 4 spaces. Might as well given all the other changes
  made.
- Separated Events into Events and Resolvers.
  - Events still encapsulate modifications to board state, but that is
    all they do, they no longer create other Events. As atomic changes,
    events record what happens to the board, which is used by both
    client and server.
  - Resolvers handle how the Events are put together, i.e. they
    encapsulate server logic for card effects and interactions, and they
    modify the board state using Events. Only the server deals with
    resolver logic, clients see changes as incoming Events. The Resolver
    structure allows for complex Event ordering and resolution,
    specifically it allows resolver nesting in place of the old flat
    event resolution.
- Optimized UI Text to render to an image buffer once per text update,
  instead of re-rendering every frame.
- Adjusted AI logic to play better.
  - Now the AI better understand the importance of mana ramping.
  - The AI previously behaved erratically when RNG and potential lethal
    was involved, now it has been made more stable.
- Revamped the Event animation system.
  - Animations can handle showing things before and after event
    resolution, and can be shown concurrently.
  - Facets about Event creation have been abstracted away in an
    EventAnimationFactory.java class.
  - A new animation has been added for EventSetEffectStats.