undo functional resolver hooks, implement conditional hook resolving
Example: a card has 2 end of turn effects, the first one destroys
itself, and the second one does something. At the end of turn, after the
first one resolves, the card is destroyed, and the second automatically
sees that the card is no longer in play, so it doesn't do anything.
Same for minions, if it has 2 on attack effects, the first one returns
the enemy minion to hand, and the second one damages it, then only the
first one triggers, and the second one sees that the minion is no longer
in play, so it doesn't attempt to damage it